Test: * Update test/changelist_data/provider.py * Update test/changelist_data/test_changelist.py * Update test/changelist_data/xml/changelists/test_changelists_reader.py * Update test/changelist_data/xml/changelists/test_changelists_tree.py * Update test/changelist_data/xml/workspace/test_workspace_reader.py * Update test/changelist_data/xml/workspace/test_workspace_tree.py
