+++ test/changelist_data/test_changelist.py
@@ -1,31 +1,10 @@
 """Testing ChangeList Data
 """
-from changelist_data.file_change import FileChange
-from changelist_data.changelist import Changelist
-
-
-def get_cl0():
-    return Changelist(
-        id="0",
-        name="",
-        changes=list(),
-    )
-
-def get_cl1():
-    return Changelist(
-        id="1212434",
-        name="ChangeList",
-        changes=[
-            FileChange(
-                after_path="/module/file.txt",
-                after_dir=False,
-            )
-        ],
-    )
+from test.changelist_data import provider
 
 
 def test_properties_cl0():
-    instance = get_cl0()
+    instance = provider.get_cl0()
     assert instance.id == '0'
     assert instance.name == ''
     assert len(instance.changes) == 0
@@ -34,7 +13,7 @@ def test_properties_cl0():
 
 
 def test_properties_cl1():
-    instance = get_cl1()
+    instance = provider.get_cl1()
     assert instance.id == '1212434'
     assert instance.name == 'ChangeList'
     assert len(instance.changes) == 1
